cleaned up the app.tsx file, all routes refactored into their own files
